Revert "Refactor worldpay controller (#36)"
This reverts commit a0cd169448ea193ce0c3fd762bbd6d76c6086a50.
