+++ spec/requests/worldpay_spec.rb
@@ -16,41 +16,104 @@ module DefraRubyMocks
         end
       end
 
-      context "#stuck" do
-        let(:path) { "/defra_ruby_mocks/worldpay/stuck" }
+      context "#payments_service" do
+        let(:path) { "/defra_ruby_mocks/worldpay/payments-service" }
 
-        context "when a direct request is made" do
-          it "renders the page without debug information and with a 200 code" do
-            get path
+        context "when a payment request is received" do
+          context "and the request is valid" do
+            let(:data) { File.read("spec/fixtures/payment_request_valid.xml") }
 
-            expect(response).to render_template(:stuck)
-            expect(response.code).to eq("200")
-            expect(response.body).to include("Stuck?")
+            it "returns an XML response with a 200 code" do
+              get path, {}, "RAW_POST_DATA" => data
+
+              expect(response.content_type).to eq("application/xml")
+              expect(response.code).to eq("200")
+              expect(response.body).to be_xml
+            end
+          end
+
+          context "and the request is invalid" do
+            let(:data) { File.read("spec/fixtures/payment_request_invalid.xml") }
+
+            it "returns a response with a 500 code" do
+              get path, {}, "RAW_POST_DATA" => data
+
+              expect(response.code).to eq("500")
+            end
           end
         end
 
-        context "when we redirect to it from the worldpay dispatcher" do
-          before(:each) do
-            cookies[:defra_ruby_mocks] = {
-              supplied_url: supplied_url,
-              separator: "?",
-              order_key: "ok",
-              mac: "mc",
-              value: 154_00,
-              status: "st",
-              reference: "rf",
-              url: "ul"
-            }.to_json
+        context "when a refund request is received" do
+          context "and the request is valid" do
+            let(:data) { File.read("spec/fixtures/refund_request_valid.xml") }
+
+            it "returns an XML response with a 200 code" do
+              get path, {}, "RAW_POST_DATA" => data
+
+              expect(response.content_type).to eq("application/xml")
+              expect(response.code).to eq("200")
+              expect(response.body).to be_xml
+            end
           end
-          let(:supplied_url) { "http://example.com/foobar" }
+        end
+      end
 
-          it "renders the page with debug information and with a 200 code" do
-            get path
+      context "#dispatcher" do
+        let(:response_url) { "#{success_url}?orderKey=admincode1^^987654&paymentStatus=#{status}&paymentAmount=10500&paymentCurrency=GBP&mac=0ba5271e1ed1b26f9bb428ef7fb536a4&source=WP" }
+        let(:path) { "/defra_ruby_mocks/worldpay/dispatcher?successURL=#{CGI.escape(success_url)}" }
+        let(:service_response) do
+          double(
+            :response,
+            supplied_url: success_url,
+            url: response_url,
+            status: status,
+            separator: "?",
+            order_key: "admincode1",
+            mac: "e5bc7ce5dfe44d2000771ac2b157f0e9",
+            value: 154_00,
+            reference: "12345"
+          )
+        end
 
-            expect(response).to render_template(:stuck)
-            expect(response.code).to eq("200")
-            expect(response.body).to include("Stuck!")
-            expect(response.body).to include(supplied_url)
+        context "and the request is valid" do
+          before(:each) { allow(WorldpayResponseService).to receive(:run) { service_response } }
+
+          let(:success_url) { "http://example.com/fo/12345/worldpay/success" }
+
+          context "and a response is expected" do
+            let(:status) { "AUTHORISED" }
+
+            it "redirects the user with a 300 code" do
+              get path
+
+              expect(response).to redirect_to(response_url)
+              expect(response.code).to eq("302")
+            end
+          end
+
+          context "and a response is not expected" do
+            let(:status) { :STUCK }
+
+            it "renders the Worldpay stuck page" do
+              get path
+
+              expect(response).to render_template(:stuck)
+              expect(response.code).to eq("200")
+            end
+          end
+        end
+
+        context "and the request is invalid" do
+          before(:each) { allow(WorldpayResponseService).to receive(:run).and_raise(MissingResourceError.new("foo")) }
+
+          context "because the success url is not in a recognised format" do
+            let(:success_url) { "http://example.com/forthewin" }
+
+            it "returns a response with a 500 code" do
+              get path
+
+              expect(response.code).to eq("500")
+            end
           end
         end
       end
@@ -59,8 +122,16 @@ module DefraRubyMocks
     context "when mocks are disabled" do
       before(:each) { DefraRubyMocks.configuration.enable = false }
 
-      context "#stuck" do
-        let(:path) { "/defra_ruby_mocks/worldpay/stuck" }
+      context "#payments_service" do
+        let(:path) { "/defra_ruby_mocks/worldpay/payments-service" }
+
+        it "cannot load the page" do
+          expect { get path }.to raise_error(ActionController::RoutingError)
+        end
+      end
+
+      context "#dispatcher" do
+        let(:path) { "/defra_ruby_mocks/worldpay/dispatcher" }
 
         it "cannot load the page" do
           expect { get path }.to raise_error(ActionController::RoutingError)
